+++ frontend/src/app/plugins/shape.cljs
@@ -399,6 +399,22 @@
                   :else
                   (st/emit! (dwsh/update-shapes [id] #(assoc % :constraints-v value))))))}
 
+           :fixedWhenScrolling
+           {:this true
+            :get #(-> % u/proxy->shape :fixed-scroll boolean)
+            :set
+            (fn [self value]
+              (cond
+                (not (boolean? value))
+                (u/not-valid plugin-id :fixedWhenScrolling value)
+
+                (not (r/check-permission plugin-id "content:write"))
+                (u/not-valid plugin-id :fixedWhenScrolling "Plugin doesn't have 'content:write' permission")
+
+                :else
+                (let [id (obj/get self "$id")]
+                  (st/emit! (dwsh/update-shapes [id] #(assoc % :fixed-scroll value))))))}
+
            :borderRadius
            {:this true
             :get #(-> % u/proxy->shape :r1)
@@ -1193,8 +1209,9 @@
                  (u/not-valid plugin-id :addGridLayout "Cannot modify a page that is not currently active")
 
                  :else
-                 (do (st/emit! (dwsl/create-layout-from-id id :grid :from-frame? true :calculate-params? false))
-                     (se/event plugin-id "create-shape-layout" :layout "grid")
+                 (do (st/emit!
+                      (dwsl/create-layout-from-id id :grid :from-frame? true :calculate-params? false)
+                      (se/event plugin-id "create-shape-layout" :layout "grid"))
                      (grid/grid-layout-proxy plugin-id file-id page-id id)))))
 
            ;; Make masks for groups