+++ frontend/test/frontend_tests/runner.cljs
@@ -5,12 +5,14 @@
    [clojure.string :as str]
    [clojure.tools.cli :refer [parse-opts]]
    [frontend-tests.basic-shapes-test]
+   [frontend-tests.code-gen-style-test]
    [frontend-tests.copy-as-svg-test]
    [frontend-tests.data.nitrate-test]
    [frontend-tests.data.repo-test]
    [frontend-tests.data.uploads-test]
    [frontend-tests.data.viewer-test]
    [frontend-tests.data.workspace-colors-test]
+   [frontend-tests.data.workspace-interactions-test]
    [frontend-tests.data.workspace-mcp-test]
    [frontend-tests.data.workspace-media-test]
    [frontend-tests.data.workspace-shortcuts-test]
@@ -26,6 +28,8 @@
    [frontend-tests.logic.pasting-in-containers-test]
    [frontend-tests.main-errors-test]
    [frontend-tests.plugins.context-shapes-test]
+   [frontend-tests.plugins.format-test]
+   [frontend-tests.plugins.interactions-test]
    [frontend-tests.plugins.page-active-validation-test]
    [frontend-tests.plugins.page-test]
    [frontend-tests.plugins.parser-test]
@@ -57,6 +61,7 @@
 
 (def test-namespaces
   '[frontend-tests.basic-shapes-test
+    frontend-tests.code-gen-style-test
     frontend-tests.copy-as-svg-test
     frontend-tests.data.nitrate-test
     frontend-tests.data.repo-test
@@ -65,6 +70,7 @@
     frontend-tests.data.uploads-test
     frontend-tests.data.viewer-test
     frontend-tests.data.workspace-colors-test
+    frontend-tests.data.workspace-interactions-test
     frontend-tests.data.workspace-mcp-test
     frontend-tests.data.workspace-media-test
     frontend-tests.data.workspace-shortcuts-test
@@ -79,6 +85,8 @@
     frontend-tests.logic.pasting-in-containers-test
     frontend-tests.plugins.context-shapes-test
     frontend-tests.plugins.page-active-validation-test
+    frontend-tests.plugins.interactions-test
+    frontend-tests.plugins.format-test
     frontend-tests.plugins.page-test
     frontend-tests.plugins.parser-test
     frontend-tests.plugins.tokens-test